requirements.txt
Faker==37.12.0
apache-airflow-providers-postgres==5.*
pandas==2.2.*
matplotlib==3.9.*
pyspark==3.5.1